Convert processor to module
